pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "giftson"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "langchain-community (>=0.3.18,<0.4.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "huggingface-hub (>=0.29.1,<0.30.0)",
    "pymupdf (>=1.25.3,<2.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
